Change x and y to be in context of sector and chunk
Make it so that instead of using world_x and world_y the program instead
uses an x or y offset from a sector which can then be combined with
chunk index and sector index to find the world_x and world_y. This
greatly simplifies a lot of the xy business.

Also removed pivot and the beginnings of the halls because there is a
lot more that needs to be done before tackling path generation.